requirements.txt
uvicorn
pydantic
openai
python-dotenv
fastapi[all]